重新设计仓储，事件等

+++ src/Ordering.Infrastructure/Migrations/20190507041915_init.cs
@@ -12,7 +12,7 @@ namespace Ordering.Infrastructure.Migrations
                 columns: table => new
                 {
                     Id = table.Column<Guid>(nullable: false),
-                    Version = table.Column<long>(nullable: false),
+                    Version = table.Column<int>(nullable: false),
                     Address_Street = table.Column<string>(nullable: true),
                     Address_City = table.Column<string>(nullable: true),
                     Address_State = table.Column<string>(nullable: true),
