+++ src/Ordering.Infrastructure/Migrations/OrderingContextModelSnapshot.cs
@@ -3,6 +3,8 @@ using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
+using Ordering.Infrastructure;
 
 namespace Ordering.Infrastructure.Migrations
 {
@@ -31,7 +33,7 @@ namespace Ordering.Infrastructure.Migrations
                     b.Property<string>("UserId")
                         .IsRequired();
 
-                    b.Property<long>("Version");
+                    b.Property<int>("Version");
 
                     b.HasKey("Id");
